Move signals module to top-level considering that viste only contains signals now anyways
